pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "cinechroma"
version = "0.1.0"
description = "Film color analysis from the command line"
readme = "README.md"
requires-python = ">=3.10"

authors = [
  { name = "Avinash", email = "avinash@example.com" }
]

dependencies = [
  "numpy",
  "opencv-python",
  "scikit-learn",
  "scikit-image",
  "pillow",
  "rich",
  "pyfiglet",
]

[project.scripts]
cinechroma = "cinechroma.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["cinechroma"]
